src/api/requirements.txt
# Kraken REST API Client Dependencies

# Core dependencies
aiohttp>=3.9.0
pydantic>=2.0.0

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0  # For environment variable management

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
mypy>=1.0.0
ruff>=0.1.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0